requirements.txt
spotipy
lyricsgenius
streamlit
pandas
transformers
torch
textblob
python-dotenv
